Create nectar scenario server boot test
Create a scenario test that boots a server. This allows us to monitor the time
taken and store it for historical monitoring purposes.

We can't use the default tempest api tests, as they create the server when
initialising the tests. This means that time taken to create the server is not
considered part of the test.

Change-Id: I7defb1af43193ec3b9cdf865d7702349de596c58