Support to specify the target host for tempest cold/live migration tests
The motivation behind such change is in live network, when we
upgrade/change compute nodes, we would like to do changes on a small portion
of nodes then verify with a bunch of tests and after that roll out to a large
scale. So it is important for us to specify src/dest host in tempest
test for this scenario.

The target_host could be left blank or once it is used, the src host should
also be specified in tempest configuration in order to avoid the src and dest
to be the same, which could fail some test cases.

Squashed with the following:
  Fixup the cold migration microversion

  This change should be squashed into change
  I7dc06929ac2f3f22883435624752b7bbabf1dc82

  Change-Id: I4fb26d31fc92b817ea051641a69ae8e3682db732
  (cherry picked from commit af77084d51820e828b79eb6a32880e87da8408ff)

Conflicts:
  tempest/api/compute/admin/test_live_migration.py

Change-Id: I7dc06929ac2f3f22883435624752b7bbabf1dc82
(cherry picked from commit c089737d9e0303dbd8d3902d68d83498e7ceb479)
(cherry picked from commit 74964d347f777e268bd5574be24f51bd7f7775cf)