Add waiter status Active or Down for ports
With OVN the port will go to DOWN status as opposed to ACTIVE
even though it's up. We also add a waiter because if too quick the
port status will be in BUILD

This is due to linuxbridge marking the ovn port as down.

Change-Id: Ia43e7bc99022341ee590f85735dd95ce98848312
(cherry picked from commit 95a641ee36043df0f0fadb6a79f890bb96629348)