Force use of newer get_project_admin_creds
get_admin_creds is not scope aware and so can return
a system scoped admin if this is set up in accounts. This
is not wanted and expects a project admin for this old call.

This is already deprecated in tempest and will be fixed in
newer versions. This just brings this forward to fix the bug
in pre prod creds.

https://bugs.launchpad.net/tempest/+bug/2072757

Change-Id: I60f9258ef31d5e3f580a017ad833290b775750a0