Nectarism: Update get_server_ip to fallback to address in server dict
Implement additional fallback to select the server IP from the
server dict if network_for_ssh is not specified and accessIPvX
attribute is empty.

Resolves an issue for nectar site-agnostic flavor tests.

For site-specific tests, our tempest config includes the option
'network_for_ssh=<site-specific-network>'. For site-agnostic
flavor tests we leave it to the nova scheduler to select
the destination site for the test server. For these
site-agnostic tests we can not specify the network_for_ssh as
network names vary across the sites. In this scenario, the
get_server_ip function queries the test server's networks for
the default ("public") which is not found.
Then the code falls back to using the accessIPv4 or accessIPv6
server attributes. However, these attributes do not get
populated consistently across all sites/default networks and
can be missing.
Hence the additional fallback to select the server IP from the
server dict.

Change-Id: Id8fdbffb6c4084a25581a7e524b4fda8998f7c8f